Moved SavepointTask to the savepoints package
